requirements.txt
pyramid_tm
pyramid_debugtoolbar
pyramid_jinja2
pyramid_beaker
sphinx
gunicorn
supervisor
zope.interface==3.8
pyramid
SQLAlchemy
transaction
zope.sqlalchemy
psycopg2
cryptacular
PyCrypto
python-dateutil==1.5
icalendar
mock
coverage
webtest

-e git+https://github.com/sontek/deform_jinja2.git#egg=deform_jinja2
-e git+https://github.com/sontek/pyramid_signup.git#egg=pyramid_signup
-e git+https://github.com/Pylons/pyramid_deform.git#egg=pyramid_deform
